backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
mistralai
google-generativeai
